ADDED: ProductDao + ProductDaoImpl + REFACTORED: Main
